Move SidebarFolderFilter to separate file
